x86: Use a simple jump into long mode
With the 64-bit descriptor we can use a jump instruction, rather than
pushing things on the stack.

Since the processor is in 64-bit mode by this point, pop a 64-bit value
from the stack, containing the target address.

This simplifies the code slightly, in particular its use of the stack.

Change-Id: Iacb56c5cbc8df98689c2a33f9d301ce4e0d746d7
Signed-off-by: Simon Glass <sjg@chromium.org>