Hack to use new Python distutils and new HOSTLDFLAGS.
Starting clang-r522817, the get_python_inc() function for distutils
no longer returns the correct value, because it is portable. This is
a hack to continue to use it to build pylibfdt.

Kleaf adds $$$$\{ORIGIN\}/... to HOSTLDFLAGS to find libc_musl.so.
This works normally for kernel builds, but it stops working when
it is used in cmd_pymod, because the $$$$, interpreted as a $$ in shell,
becomes the PID. Add a hack to fix it.

Bug: 369436075
Change-Id: I316100bff0339cfa511851b4d55c222ee9553a1d